Use react-components for open traces and views and related updates
- This change includes updates to propagate signals between webviews.
This will allow views to be updated when one view changes. For example,
when the user selects an open trace in the "Traces" view, the selection
is sent to the various react-components. The "Available Views" view
will use that signal as trigger to fetch the available views from
the remote server.

- Remove analysis and traces tree view from the extension. They are
replaced by the theia-trace-extension react components and
corresponding webviews.

- The classes AnalysisTree and TracesTree is still in the code base to
allow opening traces from the file system using the context sensitive
menu and "Open in Trace Viewer" command. This will have to be clean-up
in a later patch (+ using a cancelable progress monitor).

- The patch also includes a fix to make source-map work when debugging
the webview apps.

Needs https://github.com/theia-ide/theia-trace-extension/pull/295

Fixes #7
Fixes #11
Fixes #16

Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>