Fix bigint support in messages sent between webviews and vscode
TODO: JSONBig.parse() doesn't create bigint if numbers are small. This
has to be addressed when deserializing messages. Re-use the normalizer
of the tsp-typescript-client to do that.

The following branch in the tps-typescript-client is introducing an
utility class and, if accepted, it can be used to solve the
deserialization issue:

https://github.com/eclipse-cdt-cloud/tsp-typescript-client/pull/56

Contributes to fixing #35

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>