Fix showing of overview view with default when running in Theia
When running in Theia, the set-tspClient and set-experiment are
received right after each other and the setting of the tspClient
in the state has not been finalized since it's done asynchronously.
So, when the set-experiment signal is handled the tspClient in the
react state is still undefined and the available outputs cannot
be retrieved and hence no overview is shown.

The solution is to provide the experiment in the set-tspClient
signal if available and hence the outputs can be retrieved and
the overview view can be populated.

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>