extension: Add Start/Stop Trace Server commands
This is a WIP, to likely require introducing a back-end that may locally
or remotely hold the trace-server. The latter's path and arguments are
to be made configurable, and shall be also started upon opening traces,
if otherwise stopped. Properly stopping it on exit/quit is as required.

Include the related TODO lines and side comments as a list of WIP tasks.

Contributes to fixing (Fixes?) #15.

Signed-off-by: Marco Miller <marco.miller@ericsson.com>