Implemented external API
This commit provides an external API that can be used by other
extensions to communicate with the base trace extension. The
current API provides a way to retrieve the active experiment as well as
fire events and register callbacks. This is the initial API and will be
further expanded in future commits.

Fixes #140

Signed-off-by: Neel Gondalia <ngondalia@blackberry.com>