Automatic start trace server when user requests to open a trace
This change uses the vscode-trace-server commands to start a trace
server and when the server is started it will perform the pending
open command.

This change relies on the vscode-trace-server extension.

In a future release an API can be introduced so that an extension
can register the start/stop commands and also the started notification.
This future enhancement would decouple the vscode-trace-extension with
the vscode-trace-server implementation.

Contributes to fixing Issue #15

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>