TspClientProvider Remote Use Changes
Refactors TspClientProvider to use VSCode.asExternalUri method to generate a dynamic
uri.  This will automatically detect if we are using remote-ssh, initialize port forwarding
to the server's port, and change the provided url to a web-worker that is managing
port-forwarding to the VSCode backend.

Removes the TraceServerUrlManager class.  The functionality to generate new TspClients
when the Uri is changed is moved into this repo's TspClientProvider implementation.

Removes the tspClient.ts file.  This functionality did not work with the async nature of
asExternalUri, and was not designed to handle Uri changes.  Most functionality was already
implemented in TspClientProviderImpl.ts.

Creates VSCodeBackendTspClientProvider class.  This is a child class of TspClientProviderImpl
with expanded functionality that should only be used on the extension's backend.

Signed-off-by: Will Yang <william.yang@ericsson.com>