Use trace server manager with Open Trace command
When the command 'openTraceFolder' is invoked from the Open Trace
button, change the order so that the user is first prompted to select a
trace folder. In order to do this, the openDialog() method is exported
and is no longer called from the fileOpenHandler() method.

Use the selected trace to start the trace server using the trace server
manager, which will select the proper contributor to start the server,
based on the trace path.

Remove the execution of 'start-if-stopped' command from the
startTraceServerIfAvailable() method, and make its pathToTrace parameter
mandatory.

The fileOpenHandler() method is called last, and its traceUri parameter
is now mandatory.

Fixes #183.

Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>