[CI] Use 'list' Playwright reporter
When running the tests locally, the current 'html' reporter is fine.
But I've been confused with it in CI logs where there was a failure,
presumably a silent retry and finally a report of the tests passing.

I think in CI we would have more clarity with an alternate reporter.
Let's try 'list'. If it does not work well we can fall-back on the
simple 'dot'.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>