[CI] Have publish CI job trigger on tag creation
When we previously created the v0.2.0 release using the RELEASE file,
the corresponding GitHub release and git tag were created. However the
publish job was not triggered. Testing on my fork, I concluded that
the workflow needs to trigger on "create" tag, not on "push".

A slight side-effect of triggering the ci-cd workflow on "create" events
is that it will now run also on branch creation. So it may run a bit more
often than before. I think this is preferable for now, vs having complex
"if:" conditions on each job in the workflow.

This commit fixes this.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>