Use GitHub token from a committer in gh-release job
It turns-out that using the GitHub token, that's automatically
generated for each workflow run, does not allow for further
workflow(s) being triggered. We were relying on that feature,
for the "publish" job to trigger, after action "actions-gh-release"
creates a git tag in the repo.

Instead, for that to work, we need to use an alternate token, from
a committer. I will set such a token in the repo as a secret, under
name "GH_COMMITTER_TOKEN"

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>