Proper Handling of the URL Configuration change
This commit provides proper handling of url configuration change. This fixes the issue
of when the user changes the URL of the server, the user would need to reopen vscode so
that all of the manangers can be initialized with the proper tspClientURL. With this
commit, all tspClient dependent components are notified of a URL change.

Signed-off-by: Neel Gondalia <ngondalia@blackberry.com>