[TENTATIVE - DO NOT MERGE][Breaking] Rename extension to: eclipse-cdt.vscode-trace-viewer
Note: this is tentative for now - we will attempt to publish to the Marketplace without
changing the name at first.

From 'vscode-trace-extension'. This affects the extension's name,
used to publish it. The internal name, e.g. used in the repo's folder
name keep the original name.

This name change is necessary so we can publish the extension to the
Visual Studio Marketplace, which now requires that extensions published
there have a unique name, irrespective of publisher/namespace.

Note: this is a breaking change since other extensions that use the API
offered by this one will need to get a reference, using the vscode
extensions API, using the new name to look it up.

For example, companion extension "eclipse-cdt.vscode-trace-server",
which does that, will need to use the new name.

Follow-up: will need to adjust any reference from companion extension
vscode-trace-viewer, that uses the API. See tentative PR:
https://github.com/eclipse-cdt-cloud/vscode-trace-server/pull/31

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>