Replace PlaceholderWidget with ViewsWelcome View
Removes the PlaceholderWidget from the OpenedTracesView in the
trace-explorer.

Implements a ViewsWelcome for displaying a PlaceholderWidget equivalent.
This accurately is displayed when no traces are opened and when the
server is known to be offline.

ViewsWelcome has much less code overhead than implementing another
WebviewViewProvider or a ReactComponent from the theia-trace-extension
library.

Signed-off-by: Will Yang <william.yang@ericsson.com>