Improve Tracking of noExperiments VSCode Context
Previously this data was moving from trace-server => frontend =>
vscode-backend and was not fully flushed out.

This adds logic where the vscode-backend directly pings the trace-server
to get data on number of opened experiments, which solves problems where
the 'placeholder' or 'welcome-view' would not display or hide at the
appropriate times.

Signed-off-by: Will Yang <william.yang@ericsson.com>