Handling for Files with Open Trace Button
This commit adds a handler that manages and detects resource type
dialog that should be opened when opening a trace. It provides
the ability to handle one or both file and folder trace types.
If both is handled, a quick pick is shown to select which type of
resource dialog should be opened.

Signed-off-by: Neel Gondalia <ngondalia@blackberry.com>