Add method to set server status in TraceServerStatusService
This method set the status bar and the serverUp context that is used
to decide whether the trace explorer or the welcome view should
be rendered.

This method doesn't check the backend to get status in comparison to
checkAndUpdateServerStatus. It's useful when it's known that the
server is down or up (e.g. after start notification). In this case, it
is not necessary to check the backend again.

Don't set the status in the method isTraceServerUp() because it should
not have side-effects that are not expected.

Initialize and refresh flags traceViewer.serverUp and
trace-explorer.noExperiments to ensure the correct view is rendered,
i.e. Trace Explorer or Welcome view, when certain event happen (e.g.
server started or stopped).

Fixes #227
Fixes #228

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>