[ci] harmonize release doc, add dhelper scripts
Make this repo more similar to our other cdt-cloud repos:
- prevent triggering release-related jobs on forks
- document release process in root README
- add package.json scripts to help create the release commit and
  step version

Also, pre-format the various package.json and lerna.json files, as
per lerna expectations:
- indent using 4 spaces
- ordered dependencies

This is the format lerna enforces. Doing it now means that when we
eventually use lerna to step the package.json versions, only the
version change will be reflected in the result.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>