Enable Seperate Backend URL for Trace Server
The vscode-trace-extension code is written so that some REST calls towards
the server are done by the vscode backend, and some from the vscode frontend
(webviews). In some situations can be useful to be able to specify different
URLs for the REST calls, depending on where the REST call is done (vscode
frontend or backend). For example when deploying on a k8s cluster, where
the vscode backend can communicate to the trace server without going through
an ingress.

This commit adds the setting to be able to select the URL used from the
vscode backend.

Signed-off-by: Will Yang <william.yang@ericsson.com>
Signed-off-by: Francesco Robino <francesco.robino@ericsson.com>