perform yarn upgrade, update ovsx
The main goal was to get rid of a vulnerability in devDependency
"undici". It's being pulled by @vscode/vsce which is itself
pulled by ovsx. I updated ovsx to "0.9.4", seemingly the highest
version that still works with node 18, and after a yarn upgrade,
"undici" just barely cleared the vulnerable version range [1],
now at "6.21.1" !

[1] https://github.com/nodejs/undici/security/advisories/GHSA-c76h-2ccp-4975

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>