Add workspace file that includes trace libraries
Fixes #175

This adds a vscode/codium "multi-root"  workspace, that brings
together dependent trace libraries in source form, and the
current repo, as workspace roots. This provides a unified
development experience, where the source code of the VSCode
Trace Viewer can be handled along with the source code of
dependent trace libaries like the tsp client, timeline chart
and traceviewer base and react components libraries.

This is meant to be used as a unified development environment,
that spawns across the trace-related git repositories. It has
some limirations and a few manual setups, but already should
simplify developing features where one or more of the npm
trace libraries needs to be modified along with the code in
the repo here.

"Tasks" were also added along with the new workspace, to
permit building and watching changes in dependent libraries,
in a semi-automated way, with some limitations.

For more details, see new .md file in the repo root:

"vscode-trace-extension.code-workspace.md"

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>