Minimize jitter related to creation of trace explorer webview views
and welcome view.

Fixes #307

The initialization of webview views can be tricky. There are the 4
trace explorer views and also welcome view, one set of which should be
displayed at any given moment, depending on context variables. However
it can be possible to have jitter initially if the context variables
change value quickly, which can result in webview views being resolved
again without the "onDispose" event being sent. When this happens, the
webview views have no opportunity to unregister their event listeners,
causing leaks.

To address the jitter, context variables 'trace-explorer.noExperiments'
and 'traceViewer.serverUp' are initialized at activation with values
that will prevent the trace explorer webview views being "resolved",
causing an unambiguous situation where the welcome view will be displayed
instead, until a trace server is started or detected to be running, and
through communication with it, it's established that one or more
experiment exists on it.

In case there might still be jitter in cases that were missed, or in
case some in re-introduced in later changes, we catch the situation
where a trace explorer webview view is re-resolved without having
been disposed and in that case dispose of its listeners and print a
warning message on the console.

Also, corrected that one of the trace explorer webview views,
TraceExplorerItemPropertiesProvider, was not calling "super.dispose()"
in its own dispose() method. Since this is where the "resolved" webview
is set, this would trip the warning added in this commit, even if the
old view had been disposed correctly.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>