Properly handle on file close events
VSCode does not allow developers to have a simple 'onThisFileClose' or
'onAnyFileClose' event. So we had to use a different event to trigger
the 'onClose' event handler for customized view configuration.

The one that was being used `onDidChangeVisibleTextEditors` was not
working as expected.  The event handler has been changed to
`onDidChangeTabs`.

Fixes https://github.com/eclipse-cdt-cloud/vscode-trace-extension/issues/334

Signed-off-by: Will Yang <william.yang@ericsson.com>