Fix submitting new selection ranges from the Time Range Data View
Since commit d857eae where the vscode-messenger was introduced this
feature has been broken. The request to update the selection range
after submitting the new range from the Time Range Data view was not
propagated to the trace viewer webview where the time axis and selection
is updated. This PR fixes this.

Note that this fix sends the selection range to the correct widget and
handles the bigint values in the signal payload correctly.

fixes #357

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>